+++ CHANGELOG.md
@@ -451,6 +451,10 @@ Thanks to the people who made this release happen!
 * The revset functions `author_date()` and `committer_date()` have been renamed to
   `author_timestamp()` and `committer_timestamp()`.
 
+* The `ui.bookmark-list-sort-keys` and `ui.tag-list-sort-keys` settings now
+  accept the values `author-timestamp` and `committer-timestamp` instead of
+  `author-date` and `committer-date`.
+
 ### New features
 
 * `jj git fetch` now shows details of abandoned commits (change IDs and